--- a/db/connection.py
+++ b/db/connection.py
@@ -18,7 +18,7 @@ def execute_msar_func_with_engine(engine, func_name, *args):
     with psycopg.connect(conn_str) as conn:
         # Returns a cursor
         return conn.execute(
-            f"SELECT msar.{func_name}({','.join(['%s']*len(args))})",
+            f"SELECT msar.{func_name}({','.join(['%s'] * len(args))})",
             args
         )
 
@@ -52,7 +52,7 @@ def exec_msar_func(conn, func_name, *args):
     """
     # Returns a cursor
     return conn.execute(
-        f"SELECT msar.{func_name}({','.join(['%s']*len(args))})", args
+        f"SELECT msar.{func_name}({','.join(['%s'] * len(args))})", args
     )
 
 
@@ -66,14 +66,12 @@ def select_from_msar_func(conn, func_name, *args):
         *args: The list of parameters to pass
     """
     cursor = conn.execute(
-        f"SELECT * FROM msar.{func_name}({','.join(['%s']*len(args))})", args
+        f"SELECT * FROM msar.{func_name}({','.join(['%s'] * len(args))})", args
     )
     cursor.row_factory = dict_row
     return cursor.fetchall()
 
 
-def load_file_with_engine(engine, file_handle):
+def load_file_with_conn(conn, file_handle):
     """Run an SQL script from a file, using psycopg."""
-    conn_str = str(engine.url)
-    with psycopg.connect(conn_str) as conn:
-        conn.execute(file_handle.read())
+    conn.execute(file_handle.read())