--- a/db/tests/types/test_install.py
+++ b/db/tests/types/test_install.py
@@ -12,13 +12,6 @@ def test_create_type_schema(engine):
     assert TYPES_SCHEMA in schemata
 
 
-def test_create_type_schema_when_exists(engine):
-    # This just checks that the function doesn't error if the type schema
-    # already exists when it's run.
-    install.create_type_schema(engine)
-    install.create_type_schema(engine)
-
-
 def test_create_email_when_exists(engine):
     # This just checks that the function doesn't error if the type email
     # already exists when it's run.