--- a/db/tests/types/test_install.py
+++ b/db/tests/types/test_install.py
@@ -1,15 +1,4 @@
-from sqlalchemy import text
-from db.constants import TYPES_SCHEMA
-from db.types import install
 from db.types.custom import email
-
-
-def test_create_type_schema(engine):
-    install.create_type_schema(engine)
-    with engine.connect() as conn:
-        res = conn.execute(text("SELECT * FROM information_schema.schemata"))
-    schemata = {row['schema_name'] for row in res.fetchall()}
-    assert TYPES_SCHEMA in schemata
 
 
 def test_create_email_when_exists(engine):