--- a/db/types/install.py
+++ b/db/types/install.py
@@ -1,25 +1,5 @@
-from db.types.custom import email, money, multicurrency, uri, json_array, json_object
 from db.constants import TYPES_SCHEMA
-from db.types.operations.cast import install_all_casts
 import psycopg
-
-
-def create_type_schema(engine) -> None:
-    conn_str = str(engine.url)
-    with psycopg.connect(conn_str) as conn:
-        conn.execute(f"CREATE SCHEMA IF NOT EXISTS {TYPES_SCHEMA}")
-
-
-def install_mathesar_on_database(engine):
-    create_type_schema(engine)
-    email.install(engine)
-    money.install(engine)
-    multicurrency.install(engine)
-    uri.install(engine)
-    uri.install_tld_lookup_table(engine)
-    json_array.install(engine)
-    json_object.install(engine)
-    install_all_casts(engine)
 
 
 def uninstall_mathesar_from_database(engine):